Разделяет код в mainy.js на отдельные модули
